refactored navbar and routes to match current folder structure; added navbar and favicon icon
