+++ src/app/app.module.ts
@@ -1,14 +1,14 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { appRoutes } from '../routes';
+import { appRoutes } from '../app/routes/routes';
 
 import { AppComponent } from './app.component';
-import { NavbarComponent } from './pages/template/navbar/navbar.component';
-import { DashboardComponent } from './pages/tiles/dashboard/dashboard.component';
-import { CalendarComponent } from './pages/tiles/calendar/calendar.component';
-import { CurriculumEditorComponent } from './pages/tiles/curriculum-editor/curriculum-editor.component';
-import { BoomComponent } from './pages/tiles/boom/boom.component';
+import { NavbarComponent } from './routes/template/navbar/navbar.component';
+import { DashboardComponent } from './routes/pages/dashboard/dashboard.component';
+import { CalendarComponent } from './routes/pages/calendar/calendar.component';
+import { CurriculumEditorComponent } from './routes/pages/curriculum-editor/curriculum-editor.component';
+import { BoomComponent } from './routes/pages/boom/boom.component';
 
 @NgModule({
   declarations: [
