+++ src/app/routes/pages/boom/boom.component.ts
