+++ src/app/routes/pages/curriculum-editor/curriculum-editor.component.spec.ts
