+++ src/app/routes/pages/curriculum-editor/curriculum-editor.component.ts
