+++ src/app/routes/routes.ts
@@ -1,9 +1,9 @@
 import { Routes } from '@angular/router';
-import { AppComponent } from './app/app.component';
-import { DashboardComponent } from './app/pages/tiles/dashboard/dashboard.component';
-import { CalendarComponent } from './app/pages/tiles/calendar/calendar.component';
-import { CurriculumEditorComponent } from './app/pages/tiles/curriculum-editor/curriculum-editor.component';
-import { BoomComponent } from './app/pages/tiles/boom/boom.component';
+import { AppComponent } from '../../app/app.component';
+import { DashboardComponent } from '../../app/routes/pages/dashboard/dashboard.component';
+import { CalendarComponent } from '../../app/routes/pages/calendar/calendar.component';
+import { CurriculumEditorComponent } from '../../app/routes/pages/curriculum-editor/curriculum-editor.component';
+import { BoomComponent } from '../../app/routes/pages/boom/boom.component';
 
 export const appRoutes: Routes = [
     // routes go here.
