+++ src/app/routes/template/navbar/navbar.component.html
@@ -1,6 +1,8 @@
 <nav class="navbar navbar-expand-lg navbar-light bg-light">
 
-  <a class="navbar-brand" routerLink="/">BAM</a>
+  <a class="navbar-brand" routerLink="/">
+    <img class="icon" src="favicon.png">
+  </a>
 
   <button #hamburger class="navbar-toggler dropdown" type="button" data-toggle="collapse" data-target="#navbar" aria-controls="navbar" aria-expanded="false" aria-label="Toggle navigation">
     <span class="navbar-toggler-icon"></span>
